+++ src/azure-cli-core/azure/cli/core/__init__.py
@@ -4,7 +4,7 @@
 # --------------------------------------------------------------------------------------------
 from __future__ import print_function
 
-__version__ = "2.0.60"
+__version__ = "2.0.61"
 
 import os
 import sys
@@ -19,7 +19,6 @@ from knack.introspection import extract_args_from_signature, extract_full_summar
 from knack.log import get_logger
 from knack.util import CLIError
 from knack.arguments import ArgumentsContext  # pylint: disable=unused-import
-
 
 logger = get_logger(__name__)