+++ src/command_modules/azure-cli-rdbms/azure/cli/command_modules/rdbms/custom.py
@@ -228,6 +228,12 @@ def _replica_create(cmd, client, resource_group_name, server_name, source_server
             sku=mysql.models.Sku(name=source_server_object.sku.name),
             properties=mysql.models.ServerPropertiesForReplica(source_server_id=source_server),
             location=source_server_object.location)
+    elif provider == 'Microsoft.DBforPostgreSQL':
+        from azure.mgmt.rdbms import postgresql
+        parameters = postgresql.models.ServerForCreate(
+            sku=postgresql.models.Sku(name=source_server_object.sku.name),
+            properties=postgresql.models.ServerPropertiesForReplica(source_server_id=source_server),
+            location=source_server_object.location)
 
     return sdk_no_wait(no_wait, client.create, resource_group_name, server_name, parameters)
 
@@ -258,7 +264,6 @@ def _server_update_custom_func(instance,
                                administrator_login_password=None,
                                ssl_enforcement=None,
                                tags=None):
-    from azure.mgmt.rdbms.mysql.models import StorageProfile  # pylint: disable=unused-variable
     from importlib import import_module
     server_module_path = instance.__module__
     module = import_module(server_module_path.replace('server', 'server_update_parameters'))