+++ src/command_modules/azure-cli-kusto/HISTORY.rst
@@ -2,6 +2,13 @@
 
 Release History
 ===============
+0.2.2
++++++
+Minor fixes.
+
+0.2.1
++++++
+* kusto dependency bug fix
 
 0.2.0
 +++++