+++ src/command_modules/azure-cli-rdbms/HISTORY.rst
@@ -2,6 +2,11 @@
 
 Release History
 ===============
+
+0.3.9
++++++
+* Add postgresql replica commands and restart server command
+
 0.3.8
 +++++
 * Minor fixes: Get default location from resource group when not provided for creating servers and add validation for retention days.